Break out Command enum to its own file
