--- src/main/java/Pair.java
@@ -1,9 +1,0 @@
-public class Pair<V0, V1> {
-    V0 v0;
-    V1 v1;
-
-    Pair(V0 v0, V1 v1) {
-        this.v0 = v0;
-        this.v1 = v1;
-    }
-}
