+++ src/main/java/common/Pair.java
@@ -1,0 +1,11 @@
+package common;
+
+public class Pair<V0, V1> {
+    public V0 v0;
+    public V1 v1;
+
+    public Pair(V0 v0, V1 v1) {
+        this.v0 = v0;
+        this.v1 = v1;
+    }
+}
